+++ src/extension.ts
@@ -14,7 +14,9 @@ let _logs: Logs | null = null;
  */
 export function activate(context: vscode.ExtensionContext) {
 	const logs = _logs = new Logs(context.extension.packageJSON?.displayName);
-	logs.info('extension is now active!');
+	const platform = process.platform;  // 'darwin' / 'win32' / 'linux'
+	const arch = process.arch;          // 'x64' / 'arm64'
+	logs.log(`extension is now active! Node.js:${process.version}, VSCode:${vscode.version}, Platform:${platform}-${arch}`);
 
 	// 初期化するコマンドの登録
 	const disposable = vscode.commands.registerCommand('vscode-code-relationship-diagram.initialize', async () => {
@@ -24,18 +26,19 @@ export function activate(context: vscode.ExtensionContext) {
 		const root_folder = selectRootFolder(workspace_folders);
 		const associations = readConfiguration(vscode.workspace.getConfiguration("files").get<object>("associations"));
 		if (root_folder && (Object.keys(associations).length > 0)) {
+			const db_file = path.join(root_folder.uri.fsPath, '.vscode', 'crd.duckdb');
 			try {
 				// コード関係図DBを作成する
-				const db = new codeDb.Db(path.join(root_folder.uri.path, '.vscode', 'crd.duckdb'));
+				const db = new codeDb.Db(db_file);
 				await db.table_create();
 				try {
 					const start = performance.now();
 
 					// コードファイルを列挙する
 					const files: codeFile.File[] = [];
-					const patterns = codeFile.list(root_folder.uri.path, associations, (file: codeFile.File) => {
+					const patterns = codeFile.list(root_folder.uri.fsPath, associations, (file: codeFile.File) => {
 						files.push(file);
-						logs.info(`${(performance.now() - start).toFixed(2)} ms: list ${file.relative_path}`);
+						logs.log(`${(performance.now() - start).toFixed(2)} ms: list ${file.relative_path}`);
 					});
 
 					// コードファイルをパスでソートする
@@ -49,7 +52,7 @@ export function activate(context: vscode.ExtensionContext) {
 					for (const upsert of upserts) {
 						try {
 							await db.codeFile_upsert(upsert.relative_path, upsert.updated);
-							logs.info(`${(performance.now() - start).toFixed(2)} ms: upsert ${upsert.relative_path}`);
+							logs.log(`${(performance.now() - start).toFixed(2)} ms: upsert ${upsert.relative_path}`);
 						} catch (error) {
 							logs.trace(`db.codeFile_upsert(${upsert.relative_path}): ${error instanceof Error ? error.message : error}`);
 						}
@@ -59,21 +62,21 @@ export function activate(context: vscode.ExtensionContext) {
 					for (const remove of removes) {
 						try {
 							await db.codeFile_remove(remove);
-							logs.info(`${(performance.now() - start).toFixed(2)} ms: remove ${remove}`);
+							logs.log(`${(performance.now() - start).toFixed(2)} ms: remove ${remove}`);
 						} catch (error) {
 							logs.trace(`db.codeFile_remove(${remove}): ${error instanceof Error ? error.message : error}`);
 						}
 					}
 
-					logs.info(`${(performance.now() - start).toFixed(2)} ms`);
+					logs.log(`${(performance.now() - start).toFixed(2)} ms`);
 
 					// 初期化メッセージを表示する
 					logs.info(locale('initialize-message'));
 				} catch (error) {
-					logs.trace(`codeFile.list(${root_folder.uri.path}): ${error instanceof Error ? error.message : error}`);
+					logs.trace(`codeFile.list(${root_folder.uri.fsPath}): ${error instanceof Error ? error.message : error}`);
 				}
 			} catch (error) {
-				logs.trace(`db.table_create(): ${error instanceof Error ? error.message : error}`);
+				logs.trace(`db.table_create(${db_file}): ${error instanceof Error ? error.message : error}`);
 			}
 		} else {
 			logs.error(locale('error-no-associations'));
@@ -91,7 +94,7 @@ function selectRootFolder(folders: readonly vscode.WorkspaceFolder[]): vscode.Wo
 	// フォルダが1つ以上在って
 	return (folders.length > 0)
 		// パスの長さが最小のフォルダのパス名を返す
-		? folders.reduce((min, current) => current.uri.path.length < min.uri.path.length ? current : min)
+		? folders.reduce((min, current) => current.uri.fsPath.length < min.uri.fsPath.length ? current : min)
 		: null;
 }